simple modularization
