+++ lib/webidl2.js
@@ -1,177 +1,7 @@
 "use strict";
 
 import { syntaxError } from "./error.js";
-
-// These regular expressions use the sticky flag so they will only match at
-// the current location (ie. the offset of lastIndex).
-const tokenRe = {
-  // This expression uses a lookahead assertion to catch false matches
-  // against integers early.
-  "decimal": /-?(?=[0-9]*\.|[0-9]+[eE])(([0-9]+\.[0-9]*|[0-9]*\.[0-9]+)([Ee][-+]?[0-9]+)?|[0-9]+[Ee][-+]?[0-9]+)/y,
-  "integer": /-?(0([Xx][0-9A-Fa-f]+|[0-7]*)|[1-9][0-9]*)/y,
-  "identifier": /[_-]?[A-Za-z][0-9A-Z_a-z-]*/y,
-  "string": /"[^"]*"/y,
-  "whitespace": /[\t\n\r ]+/y,
-  "comment": /((\/(\/.*|\*([^*]|\*[^/])*\*\/)[\t\n\r ]*)+)/y,
-  "other": /[^\t\n\r 0-9A-Za-z]/y
-};
-
-const stringTypes = [
-  "ByteString",
-  "DOMString",
-  "USVString"
-];
-
-const argumentNameKeywords = [
-  "attribute",
-  "callback",
-  "const",
-  "deleter",
-  "dictionary",
-  "enum",
-  "getter",
-  "includes",
-  "inherit",
-  "interface",
-  "iterable",
-  "maplike",
-  "namespace",
-  "partial",
-  "required",
-  "setlike",
-  "setter",
-  "static",
-  "stringifier",
-  "typedef",
-  "unrestricted"
-];
-
-const nonRegexTerminals = [
-  "-Infinity",
-  "FrozenArray",
-  "Infinity",
-  "NaN",
-  "Promise",
-  "boolean",
-  "byte",
-  "double",
-  "false",
-  "float",
-  "implements",
-  "legacyiterable",
-  "long",
-  "mixin",
-  "null",
-  "octet",
-  "optional",
-  "or",
-  "readonly",
-  "record",
-  "sequence",
-  "short",
-  "true",
-  "unsigned",
-  "void"
-].concat(argumentNameKeywords, stringTypes);
-
-const punctuations = [
-  "(",
-  ")",
-  ",",
-  "...",
-  ":",
-  ";",
-  "<",
-  "=",
-  ">",
-  "?",
-  "[",
-  "]",
-  "{",
-  "}"
-];
-
-function tokenise(str) {
-  const tokens = [];
-  let lastCharIndex = 0;
-  let trivia = "";
-  let line = 1;
-  let index = 0;
-  while (lastCharIndex < str.length) {
-    const nextChar = str.charAt(lastCharIndex);
-    let result = -1;
-
-    if (/[\t\n\r ]/.test(nextChar)) {
-      result = attemptTokenMatch("whitespace", { noFlushTrivia: true });
-    } else if (nextChar === '/') {
-      result = attemptTokenMatch("comment", { noFlushTrivia: true });
-    }
-
-    if (result !== -1) {
-      const currentTrivia = tokens.pop().value;
-      line += (currentTrivia.match(/\n/g) || []).length;
-      trivia += currentTrivia;
-      index -= 1;
-    } else if (/[-0-9.A-Z_a-z]/.test(nextChar)) {
-      result = attemptTokenMatch("decimal");
-      if (result === -1) {
-        result = attemptTokenMatch("integer");
-      }
-      if (result === -1) {
-        result = attemptTokenMatch("identifier");
-        const token = tokens[tokens.length - 1];
-        if (result !== -1 && nonRegexTerminals.includes(token.value)) {
-          token.type = token.value;
-        }
-      }
-    } else if (nextChar === '"') {
-      result = attemptTokenMatch("string");
-    }
-
-    for (const punctuation of punctuations) {
-      if (str.startsWith(punctuation, lastCharIndex)) {
-        tokens.push({ type: punctuation, value: punctuation, trivia, line, index });
-        trivia = "";
-        lastCharIndex += punctuation.length;
-        result = lastCharIndex;
-        break;
-      }
-    }
-
-    // other as the last try
-    if (result === -1) {
-      result = attemptTokenMatch("other");
-    }
-    if (result === -1) {
-      throw new Error("Token stream not progressing");
-    }
-    lastCharIndex = result;
-    index += 1;
-  }
-
-  // remaining trivia as eof
-  tokens.push({
-    type: "eof",
-    value: "",
-    trivia
-  });
-
-  return tokens;
-
-  function attemptTokenMatch(type, { noFlushTrivia } = {}) {
-    const re = tokenRe[type];
-    re.lastIndex = lastCharIndex;
-    const result = re.exec(str);
-    if (result) {
-      tokens.push({ type, value: result[0], trivia, line, index });
-      if (!noFlushTrivia) {
-        trivia = "";
-      }
-      return re.lastIndex;
-    }
-    return -1;
-  }
-}
+import { argumentNameKeywords, stringTypes, tokenise } from "./tokeniser.js";
 
 class WebIDLParseError extends Error {
   constructor({ message, line, input, tokens }) {
